+++ sdk/tables/azure-data-tables/azure/data/tables/_table_client.py
@@ -22,6 +22,7 @@ from ._error import _process_table_error, _validate_table_name
 from ._generated.models import (
     SignedIdentifier,
     TableProperties,
+    QueryOptions
 )
 from ._serialize import _get_match_headers, _add_entity_properties
 from ._base_client import parse_connection_str, TablesBaseClient
@@ -439,8 +440,8 @@ class TableClient(TablesBaseClient):
         # type: (...) -> ItemPaged[TableEntity]
         """Lists entities in a table.
 
-        :keyword int results_per_page: Number of entities per page in return ItemPaged
-        :keyword select: Specify desired properties of an entity to return certain entities
+        :keyword int results_per_page: Number of entities returned per service request.
+        :keyword select: Specify desired properties of an entity to return.
         :paramtype select: str or List[str]
         :return: ItemPaged[:class:`~azure.data.tables.TableEntity`]
         :rtype: ~azure.core.paging.ItemPaged
@@ -479,8 +480,8 @@ class TableClient(TablesBaseClient):
         """Lists entities in a table.
 
         :param str query_filter: Specify a filter to return certain entities
-        :keyword int results_per_page: Number of entities per page in return ItemPaged
-        :keyword select: Specify desired properties of an entity to return certain entities
+        :keyword int results_per_page: Number of entities returned per service request.
+        :keyword select: Specify desired properties of an entity to return.
         :paramtype select: str or List[str]
         :keyword Dict[str, Any] parameters: Dictionary for formatting query with additional, user defined parameters
         :return: ItemPaged[:class:`~azure.data.tables.TableEntity`]
@@ -503,7 +504,7 @@ class TableClient(TablesBaseClient):
         top = kwargs.pop("results_per_page", None)
         user_select = kwargs.pop("select", None)
         if user_select and not isinstance(user_select, str):
-            user_select = ", ".join(user_select)
+            user_select = ",".join(user_select)
 
         command = functools.partial(self._client.table.query_entities, **kwargs)
         return ItemPaged(
@@ -529,6 +530,8 @@ class TableClient(TablesBaseClient):
         :type partition_key: str
         :param row_key: The row key of the entity.
         :type row_key: str
+        :keyword select: Specify desired properties of an entity to return.
+        :paramtype select: str or List[str]
         :return: Dictionary mapping operation metadata returned from the service
         :rtype: :class:`~azure.data.tables.TableEntity`
         :raises: :class:`~azure.core.exceptions.HttpResponseError`
@@ -542,14 +545,17 @@ class TableClient(TablesBaseClient):
                 :dedent: 8
                 :caption: Get a single entity from a table
         """
+        user_select = kwargs.pop("select", None)
+        if user_select and not isinstance(user_select, str):
+            user_select = ",".join(user_select)
         try:
             entity = self._client.table.query_entity_with_partition_and_row_key(
                 table=self.table_name,
                 partition_key=partition_key,
                 row_key=row_key,
+                query_options=QueryOptions(select=user_select),
                 **kwargs
             )
-
             properties = _convert_to_entity(entity)
             return properties
         except HttpResponseError as error: