+++ sdk/tables/azure-data-tables/CHANGELOG.md
@@ -21,6 +21,7 @@
 * Added support for Azurite storage emulator
 * Throws a `RequestTooLargeError` on transaction requests that return a 413 error code
 * Added support for Int64 and Binary types in query filters
+* Added support for `select` keyword parameter to `TableClient.get_entity()`.
 
 ## 12.0.0b6 (2021-04-06)
 * Updated deserialization of datetime fields in entities to support preservation of the service format with additional decimal place.